% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtrack.r
\name{backtrack}
\alias{backtrack}
\title{Backtracking Line Search}
\usage{
backtrack(x, dx, f, df, alpha = 0.01, beta = 0.8)
}
\arguments{
\item{x}{Current iterate}

\item{dx}{Descent direction}

\item{f}{objective function}

\item{df}{gradient of objective function}

\item{alpha}{sufficient decrease parameter}

\item{beta}{sufficient decrease parameter}
}
\description{
Given a descent direction \code{backtrack} computes a step size that ensures sufficient decrease in an objective.
}

