% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg.r
\name{dg}
\alias{dg}
\title{Compute the gradient of the majorization.}
\usage{
dg(x, xa, v, w, plist1, plist2, h, hgrad)
}
\arguments{
\item{x}{non-anchor point}

\item{xa}{Anchor point}

\item{v}{weights for first set of constraints}

\item{w}{weights for second set of constraints}

\item{plist1}{list of projection functions for first set of constraints; each takes a single point and returns its projection}

\item{plist2}{list of projection functions for second set of constraints; each takes a single point and returns its projection}

\item{h}{Function handle for output mapping}

\item{hgrad}{Handle for output mapping Jacobian}
}
\description{
\code{dg} computes the gradient of the majorization of the proximity function.
}

