% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmqn_step.r
\name{mmqn_step}
\alias{mmqn_step}
\title{MM-quasi-Newton step}
\usage{
mmqn_step(x, v, w, plist1, plist2, f, df, h, hgrad, woodbury = TRUE)
}
\arguments{
\item{x}{Current iterate}

\item{v}{weights for first set of constraints}

\item{w}{weights for second set of constraints}

\item{plist1}{list of projection functions for first set of constraints; each takes a single point and returns its projection}

\item{plist2}{list of projection functions for second set of constraints; each takes a single point and returns its projection}

\item{f}{objective function}

\item{df}{gradient of objective function}

\item{h}{Function handle for output mapping}

\item{hgrad}{Handle for output mapping Jacobian}

\item{woodbury}{Boolean: TRUE to use the Woodbury inversion formula}
}
\description{
\code{mmqn_step} computes a single step.
}

