% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmsfp_sap_one_step.r
\name{nmsfp_sap_one_step}
\alias{nmsfp_sap_one_step}
\title{One step of self-adaptive projection-type method for the NMSFP}
\usage{
nmsfp_sap_one_step(x, delta, mu, tau, gamma, df, proj)
}
\arguments{
\item{x}{current iterate}

\item{delta}{step-size parameter}

\item{mu}{step-size parameter}

\item{tau}{step-size parameter}

\item{gamma}{step-size parameter}

\item{df}{handle to gradient of objective function}

\item{proj}{handle to projection operation.}
}
\description{
\code{nmsfp_sap_one_step} performs the self-adaptive projection-type method of Li et al.
}
\references{
Li, Han, and Zhang. (2013) A self-adaptive projection-type method for nonlinear 
multiple-sets split feasibility problem, Inverse Problems in Science and Engineering,
21(1):155-170.
}

