% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{project_square}
\alias{project_square}
\title{Project onto a square}
\usage{
project_square(x, center, r)
}
\arguments{
\item{x}{Point to project}

\item{center}{Center of the square}

\item{r}{Half the length of a side}
}
\description{
\code{project_square} computes the Euclidean projection of a point onto a square.
}
\examples{
set.seed(12345)
p <- 2
center <- matrix(rnorm(p),p,1)
r <- runif(1)
x <- matrix(rnorm(p),p,1)
y <- project_square(x,center,r)
}
