% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.r
\name{proximity}
\alias{proximity}
\title{Proximity function}
\usage{
proximity(x, v, w, plist1, plist2, h)
}
\arguments{
\item{x}{Current iterate}

\item{v}{weights for first set of constraints}

\item{w}{weights for second set of constraints}

\item{plist1}{list of projection functions for first set of constraints; each takes a single point and returns its projection}

\item{plist2}{list of projection functions for second set of constraints; each takes a single point and returns its projection}

\item{h}{Function handle for output mapping}
}
\description{
\code{proximity} computes the proximity function.
}

