% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnamm.r
\name{qnamm}
\alias{qnamm}
\title{Quasi-Newton acceleration of MM algorithm}
\usage{
qnamm(x, fx_mm, qn, fx_obj, max_iter = 100, tol = 1e-06, ...)
}
\arguments{
\item{x}{initial iterate}

\item{fx_mm}{MM algorithm map}

\item{qn}{number of secants}

\item{fx_obj}{handle to objective function}

\item{max_iter}{maximum number of iterations}

\item{tol}{convergence tolerance}

\item{...}{Additional arguments to pass to \code{fx_mm}}
}
\description{
\code{qnamm} performs Quasi-Newton acceleration of an MM algorithm.
}
\references{
H Zhou, D Alexander, and K Lange. (2011) A quasi-Newton acceleration method for high-dimensional optimization algorithms, Statistics and Computing, 21(2):261-273.
}

