% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.r
\name{softmax}
\alias{softmax}
\title{Compute soft-max}
\usage{
softmax(x, a = 100)
}
\arguments{
\item{x}{input}

\item{a}{scaling factor}
}
\description{
\code{softmax} returns the soft maximum of a collection of reals.
}
\examples{
set.seed(12345)
n <- 10
x <- rnorm(n)
softmax(x)
}

