% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_ball.r
\name{project_ball}
\alias{project_ball}
\title{Projection onto a ball}
\usage{
project_ball(x, cnt, r)
}
\arguments{
\item{x}{Point to project}

\item{r}{Radius of the sphere}

\item{center}{Center of the sphere}
}
\description{
\code{project_ball} computes the Euclidean projection of a point onto a ball.
}
\examples{
set.seed(12345)
p <- 3
cnt <- rnorm(p)
r <- runif(1)
x <- rnorm(p)
y <- project_ball(x,cnt,r)
}

