% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{project_cube}
\alias{project_cube}
\title{Project onto a cube}
\usage{
project_cube(x, center, r)
}
\arguments{
\item{x}{Point to project}

\item{center}{Center of the square}

\item{r}{Half the length of a side}
}
\description{
\code{project_cube} computes the Euclidean projection of a point onto a cube.
}
\examples{
set.seed(12345)
p <- 3
center <- matrix(rnorm(p),p,1)
r <- runif(1)
x <- matrix(rnorm(p),p,1)
y <- project_cube(x,center,r)
}
